requirements.txt
Django==3.2.16
django-colorfield==0.7.2
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
